fix: make the last semester credit card section to unfoldable
